Use interfaces in services instread of impls

+++ internal/service/contract.go
@@ -9,7 +9,7 @@ import (
 //go:generate mockgen --build_flags=--mod=mod -destination mock/mock_service.go . IIssueService,IProjectService
 
 type IIssueService interface {
-	GetIssuesByProject(ctx context.Context, projectId int64) ([]domain.Issue, error)
+	GetIssuesByProject(ctx context.Context, projectId int) ([]domain.Issue, error)
 }
 
 type IProjectService interface {
